REFACTOR: Use new ACSets.jl for acsets implementation.
Remove internal acsets implementation.
