+++ Project.toml
@@ -5,6 +5,7 @@ authors = ["Evan Patterson <evan@epatters.org>"]
 version = "0.14.17"
 
 [deps]
+ACSets = "227ef7b5-1206-438b-ac65-934d6da304b8"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 CompTime = "0fb5dd42-039a-4ca4-a1d7-89a96eae6d39"
 Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
@@ -15,7 +16,6 @@ LightXML = "9c8b4983-aa76-5018-a973-4c85ecc9e179"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 MLStyle = "d8e11817-5142-5d16-987a-aa16d5891078"
-Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
@@ -44,6 +44,7 @@ CatlabSCSExt = "SCS"
 CatlabTikzPicturesExt = "TikzPictures"
 
 [compat]
+ACSets = "0.1"
 Colors = "0.12"
 CompTime = "0.1"
 Compose = "0.7, 0.8, 0.9"
