+++ src/Catlab.jl
@@ -1,12 +1,9 @@
 module Catlab
 
-include("acsets/ACSets.jl")
-
 include("core/Core.jl")
 include("theories/Theories.jl")
 
-include("acsets/CSetDataStructures.jl")
-
+include("categorical_algebra/ACSetsGATsInterop.jl")
 include("categorical_algebra/Permutations.jl")
 include("graphs/Graphs.jl")
 include("categorical_algebra/CategoricalAlgebra.jl")
