--- src/acsets/ACSets.jl
@@ -1,14 +1,0 @@
-using Reexport 
-
-include("IndexUtils.jl")
-include("LVectors.jl")
-include("Defaults.jl")
-include("Mappings.jl")
-include("PreimageCaches.jl")
-include("Columns.jl")
-include("ColumnImplementations.jl")
-include("Schemas.jl")
-include("ACSetInterface.jl")
-include("DenseACSets.jl")
-
-@reexport using .ColumnImplementations: AttrVar
