+++ src/graphs/GraphAlgorithms.jl
@@ -6,8 +6,9 @@ export connected_components, connected_component_projection, connected_component
 
 using DataStructures: Stack, DefaultDict
 
+using ACSets
 using ...Theories: dom, codom
-using ...CSetDataStructures, ..BasicGraphs
+using ..BasicGraphs
 using ..PropertyGraphs: ReflexiveEdgePropertyGraph
 
 # Connectivity
