+++ src/theories/Theories.jl
@@ -6,10 +6,11 @@ are also included.
 module Theories
 export CategoryExpr, ObExpr, HomExpr
 
-using ..Catlab
-import ..Schemas: dom, codom, ob, hom, attr, attrtype
-import ..Syntax: GATExpr, show_unicode, show_latex
 import Base: ≤
+import ACSets: dom, codom, ob, hom, attr, attrtype
+
+using ..Catlab
+import ..Syntax: GATExpr, show_unicode, show_latex
 
 """ Base type for GAT expressions in categories or other categorical structures.
